# Improve performance by switching from `git2` to `gitoxide`