--- a/Cargo.toml
+++ b/Cargo.toml
@@ -58,8 +58,9 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_yaml = "0.9"
 sha1 = "0.10"
-tracing = "0.1"
-tracing-subscriber = "0.3"
+tracing = "0.1.37"
+tracing-log = "0.1.3"
+tracing-subscriber = { version = "0.3.16", features = ["tracing-log", "ansi"] }
 walkdir = "2.3"
 
 [dev-dependencies]