--- a/Cargo.toml
+++ b/Cargo.toml
@@ -32,10 +32,9 @@ atty = "0.2"
 bstr = { version = "1.0.1", features = ["serde"] }
 clap = { version = "4.0", features = ["cargo", "derive", "env", "unicode", "wrap_help"] }
 console = "0.15.2"
-git2 = { version = "0.15", features = ["vendored-libgit2", "vendored-openssl"] }
+git-discover = "0.10.0"
 git-repository = { version = "0.30.0", features = ["max-performance"] }
 libc = "0.2"
-libgit2-sys = "*"
 hex = "0.4"
 hyperscan = { version = "0.3", features = ["full", "static"] }
 # hyperscan-sys = { version = "0.3", features = ["full", "static"] }
@@ -64,8 +63,10 @@ tracing-subscriber = { version = "0.3.16", features = ["tracing-log", "ansi"] }
 walkdir = "2.3"
 
 [dev-dependencies]
-proptest = "1.0"
 criterion = "0.4"
+libgit2-sys = "0.14.0"
+git2 = { version = "0.15", features = ["vendored-libgit2", "vendored-openssl"] }
+proptest = "1.0"
 
 [[bench]]
 name = "microbench"