--- a/src/input_enumerator.rs
+++ b/src/input_enumerator.rs
@@ -1,5 +1,6 @@
-use anyhow::{bail, Result};
+use anyhow::{bail, Context, Result};
 use git2::{Oid, Repository, RepositoryOpenFlags};
+use git_repository as git;
 use ignore::{WalkBuilder, WalkState};
 use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
@@ -169,8 +170,7 @@ impl<'t> ignore::ParallelVisitor for Visitor<'t> {
             //
             // Had follow_symlinks been enabled, the pointed-to entry would have been yielded
             // instead.
-        }
-        else {
+        } else {
             warn!("Unhandled path type: {:?}: {:?}", path, entry.file_type());
         }
         WalkState::Continue
@@ -256,7 +256,7 @@ impl FilesystemEnumerator {
 }
 
 /// Opens the given Git repository if it exists, returning None otherwise.
-pub fn open_git_repo(path: &Path) -> Result<Option<Repository>> {
+pub fn open_git2_repo(path: &Path) -> Result<Option<Repository>> {
     match Repository::open_ext(
         path,
         RepositoryOpenFlags::NO_SEARCH | RepositoryOpenFlags::NO_DOTGIT, // | RepositoryOpenFlags::BARE,
@@ -270,38 +270,61 @@ pub fn open_git_repo(path: &Path) -> Result<Option<Repository>> {
     }
 }
 
+/// Opens the given Git repository if it exists, returning None otherwise.
+pub fn open_git_repo(path: &Path) -> Result<Option<git::Repository>> {
+    let opts = {
+        let mut opts = git::open::Options::isolated();
+        opts.permissions.env.objects = git::sec::Permission::Allow;
+        opts
+    };
+    match git::open_opts(path, opts) {
+        Err(git::open::Error::NotARepository { .. }) => Ok(None),
+        Err(err) => Err(err.into()),
+        Ok(r) => Ok(Some(r)),
+    }
+}
+
 pub struct GitRepoEnumeratorResult {
     pub blobs: Vec<(Oid, u64)>,
 }
 
 pub struct GitRepoEnumerator<'a> {
-    repo: &'a Repository,
+    repo: &'a git::Repository,
 }
 
 impl<'a> GitRepoEnumerator<'a> {
-    pub fn new(repo: &'a Repository) -> Self {
+    pub fn new(repo: &'a git::Repository) -> Self {
         GitRepoEnumerator { repo }
     }
 
     pub fn run(&self, progress: &mut Progress) -> Result<GitRepoEnumeratorResult> {
+        use git::prelude::HeaderExt;
+
         let mut blobs: Vec<(Oid, u64)> = Vec::new();
 
-        let odb = self.repo.odb()?;
-        odb.foreach(|oid: &git2::Oid| {
-            let (obj_size, obj_type) = match odb.read_header(*oid) {
-                Err(e) => {
-                    error!("Failed to read object header {}: {}", oid, e);
-                    return true;
+        let odb = &self.repo.objects;
+        for oid in odb
+            .iter()?
+            .with_ordering(
+                git::odb::store::iter::Ordering::PackAscendingOffsetThenLooseLexicographical,
+            )
+            .filter_map(Result::ok)
+        {
+            let hdr = odb
+                .header(oid)
+                .with_context(|| format!("Failed to read object header {}", oid))?;
+            let obj_type = hdr.kind();
+            match obj_type {
+                git::object::Kind::Blob => {
+                    let obj_size = hdr.size() as u64;
+                    progress.inc(obj_size);
+                    blobs.push((git2::Oid::from_bytes(oid.as_bytes())?, obj_size));
+                    // let read_size = odb.read(*oid).unwrap().len();
+                    // assert_eq!(obj_size, read_size);
                 }
-                Ok(v) => v,
-            };
-            if obj_type == git2::ObjectType::Blob {
-                let obj_size = obj_size as u64;
-                progress.inc(obj_size);
-                blobs.push((*oid, obj_size));
+                _ => {}
             }
-            true
-        })?;
+        }
 
         Ok(GitRepoEnumeratorResult { blobs })
     }