This is a revised and expanded version of PR #2, co-authored by @Byron.