@bradlarsen 2022-12-21
Some benchmark results. Baseline is `main`; candidate is the code from this PR.

2019 MBP (2.6 GHz 6-Core Intel Core i7, 64 GB 2667 MHz DDR4, 1TB NVMe).

Both were running release builds of Nosey Parker.

## Linux Kernel
This is from `git clone --mirror https://github.com/torvalds/linux`.
### Baseline
```
Found 102.02 GiB from 18 plain files and 2,829,263 blobs from 1 Git repos [00:00:33]
Scanning content  ████████████████████ 100%  102.02 GiB/102.02 GiB  [00:02:26]
Scanned 102.02 GiB from 2,829,281 blobs in 2 minutes (715.31 MiB/s); 50/50 new matches
...
900.26s user 137.86s system 575% cpu 3:00.45 total
```

### Candidate
```
Found 102.02 GiB from 18 plain files and 2,829,263 blobs from 1 Git repos [00:00:19]
Scanning content  ████████████████████ 100%  102.02 GiB/102.02 GiB  [00:01:07]
Scanned 102.02 GiB from 2,829,281 blobs in 68 seconds (1.50 GiB/s); 50/50 new matches
...
721.19s user 8.49s system 820% cpu 1:28.90 total
```

So a bit more than a 2x speedup here in real time from switching from `git2` to `gitoxide`. 🎉