@bradlarsen 2022-12-22
Another benchmark. Baseline again is `main`; candidate is the code from this PR.

This is running on a 64-core n1-standard-64 instance on Google Cloud (64 hyperthreaded cores; 236GB RAM; "balanced" persistent disk).

## 7479 Git repositories

### Baseline
```
Found 1.77 TiB from 134,849 plain files and 45,220,027 blobs from 7,479 Git repos [00:05:29]
Scanning content  ████████████████████ 100%  1.77 TiB/1.77 TiB  [00:19:36]                                                     Scanned 1.35 TiB from 27,775,304 blobs in 20 minutes (1.17 GiB/s); 996,420/996,420 new matches
...
18778.06s user 2855.37s system 1435% cpu 25:07.01 total
```

### Candidate
```
Found 1.77 TiB from 134,849 plain files and 45,220,027 blobs from 7,479 Git repos [00:00:24]
Scanning content  ████████████████████ 100%  1.77 TiB/1.77 TiB  [00:07:30]                                                    Scanned 1.36 TiB from 27,709,513 blobs in 8 minutes (3.09 GiB/s); 996,420/996,420 new matches
...
24947.43s user 1372.01s system 5404% cpu 8:06.97 total
```

In this case, using `gitoxide` instead of `git2` gives more than a 3x overall speedup in real time. The `gitoxide` version scales much better to many cores: for nearly all the runtime, all 64 cores were pegged at 100%.

In this input data, there are a handful of the 7479 repositories that cause Nosey Parker to spend lots of time doing `regex` matching rather than reading Git blobs. When I can finally address that, the speedup from using `gitoxide` should be even greater.

CC @Byron